requirements.txt
torchtext == 0.6.0
torchvision >= 0.6.0
timm == 0.3.4
fairscale